Import csv2wiki and use in simple case.
When importing, done as a vcs style package, so that running `pipenv
install` will update as things are done.  Upload is very simple and
counts on you basically being able to have run wiki-refresh in macfound
with zero errors (if the previous statement makes no sense to you, this
project has moved forward enough that it doesn't need to).