Add username, password setting, signin_required.
For standalone mode, there needs to be a way to set the username and
password for testing and development purposes.  Instead of pulling in
the full account management from librehq-core, we add a mock
configuration that only works in standalone mode.

This paves the way for doing wiki maintenance by user.