Install Mediawiki and enable in webserver
At this point, the wiki comes up in the browser when navigating to the
address given in the ServerName directive in mwiki-default.conf.
However, the user is still required to click a link and fill out
settings for the wiki.

I expect that adding a LocalSettings.php with some system defaults will
bypass this step.  Next!

Note also that running the installation script appears not to be
idempotent.  Though it has worked (as indicated in 97322f3a), it often
fails.  I need to find the correct conditional to set so that the
playbook can be run multiple times.

Note the PHP 7.2 and 7.3 versions of the XML module.  I believe that is
necessary only due to a quirk of the server I am using to test.  I
should be able to only include the 7.3 version.