Install Mediawiki using Composer
Follow @frankduncan's example in his local setup script and use Composer
to install Mediawiki and its dependencies.  Unfortunately, this still
leads to a screen appearing that asks the user to "Please set up the
wiki first."

This version of the playbook is idempotent, though, which is a marked
improvement over the last version.

I moved the symlink to the web server directory so that it occurs after
MediaWiki installation to avoid permissions issues.  Also, now all the
Apache configuration work happens together at the end.