Install MWFarm with Ansible
This now works correctly.  Keys were making sure that the operation is
as idempotent as possible (removing MWFarm before re-installing it),
paying attention to ownership, and moving the Apache changes to the end.
That last is important because giving www-data ownership over the
symlink means that the wiki user can't add new files to the `mediawiki`
subdir, which is bad!  I'd like to handle that better.