Update ansible to work on aws localhost
The current ansible playbook didn't work when pointed at localhost on a
fresh aws instance, because of placeholder variables and what looked
like a hardcoded setup to vagrant.

These changes may need to be done to pull in variables from somewhere
else during an installation, but for now they can be hard coded.